backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
python-dotenv
slack-sdk
notion-client
google-api-python-client
google-auth
google-auth-oauthlib
google-auth-httplib2
apscheduler
pytest
pytest-asyncio
sqlalchemy
alembic